requirements.txt
flask
flask-socketio
argon2-cffi
gevent
gevent-websocket
